feat: 상품 가입 구현

+++ src/main/java/com/project/hana_piece/product/dto/EnrollProductRequest.java
@@ -7,7 +7,7 @@ public record EnrollProductRequest (
         Long userGoalId,
         Integer contractPeriod,
         BigInteger initialAmount,
-        BigInteger autoDebitAmount,
+        Long autoDebitAmount,
         Integer autoDebitDay,
         String maturityDate,
         boolean autoRenewal
